[tests] split and shorten tests
