+++ tests/test_dependencies/hardpy.toml
@@ -1,0 +1,12 @@
+title = "HardPy TOML config"
+tests_dir = "tests"
+
+[database]
+user = "dev"
+password = "dev"
+host = "localhost"
+port = 5984
+
+[frontend]
+host = "localhost"
+port = 8000
