+++ tests/test_dependencies/test_error_handling.py
@@ -1,0 +1,137 @@
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from pytest import Pytester
+
+status_test_header = """
+        import pytest
+        """
+
+
+def test_case_dependency_with_incorrect_data(
+    pytester: Pytester, hardpy_opts: list[str],
+):
+    pytester.makepyfile(
+        test_1="""
+        import pytest
+        @pytest.mark.dependency(":::")
+        def test_one():
+            assert True
+    """,
+    )
+    result = pytester.runpytest(*hardpy_opts)
+    result.assert_outcomes()
+
+
+def test_module_dependency_with_incorrect_case(
+    pytester: Pytester,
+    hardpy_opts: list[str],
+):
+    pytester.makepyfile(
+        test_1="""
+        import pytest
+
+        def test_one():
+            assert True
+    """,
+    )
+    pytester.makepyfile(
+        test_2="""
+        import pytest
+
+        pytestmark = pytest.mark.dependency("test_1::fghdjshfj")
+
+        def test_one():
+            assert True
+    """,
+    )
+    result = pytester.runpytest(*hardpy_opts)
+    result.assert_outcomes()
+
+
+def test_case_dependency_with_incorrect_module_and_case(
+    pytester: Pytester,
+    hardpy_opts: list[str],
+):
+    pytester.makepyfile(
+        test_1="""
+        import pytest
+
+        @pytest.mark.dependency("kuhfkhgf::qwqw")
+        def test_one():
+            assert True
+    """,
+    )
+    result = pytester.runpytest(*hardpy_opts)
+    result.assert_outcomes()
+
+
+def test_case_dependency_with_incorrect_module(
+    pytester: Pytester,
+    hardpy_opts: list[str],
+):
+    pytester.makepyfile(
+        test_1="""
+        import pytest
+
+        @pytest.mark.dependency("kuhfkhgf")
+        def test_one():
+            assert True
+    """,
+    )
+    result = pytester.runpytest(*hardpy_opts)
+    result.assert_outcomes()
+
+
+def test_module_dependency_with_incorrect_data(
+    pytester: Pytester,
+    hardpy_opts: list[str],
+):
+    pytester.makepyfile(
+        test_1="""
+        import pytest
+
+        def test_one():
+            assert True
+    """,
+    )
+    pytester.makepyfile(
+        test_2="""
+        import pytest
+
+        pytestmark = pytest.mark.dependency(":::")
+
+        def test_one():
+            assert True
+    """,
+    )
+    result = pytester.runpytest(*hardpy_opts)
+    result.assert_outcomes()
+
+
+def test_module_dependency_with_incorrect_module_and_correct_case(
+    pytester: Pytester,
+    hardpy_opts: list[str],
+):
+    pytester.makepyfile(
+        test_1="""
+        import pytest
+
+        def test_one():
+            assert True
+    """,
+    )
+    pytester.makepyfile(
+        test_2="""
+        import pytest
+
+        pytestmark = pytest.mark.dependency("asfddaa::test_one")
+
+        def test_one():
+            assert True
+    """,
+    )
+    result = pytester.runpytest(*hardpy_opts)
+    result.assert_outcomes()
